+++ tests/wdio/config/local.ts
@@ -1,0 +1,7 @@
+import { config as baseConfig } from './base';
+
+export const config: WebdriverIO.Config = {
+  ...baseConfig,
+
+  services: [...baseConfig.services, 'chromedriver'],
+};